requirements.txt
langchain
langchain-core
langchain-community
langchain-groq
langchain-openai
langchain-chroma
ipykernel
python-dotenv
streamlit
pypdf